Add ARM support. Cleaned up the waf support.
Adding ARM support has changed the architecture relocation interface.
All architectures updated.

Cleaned up the RTEMS waf support to only look for tools when the arch
is being processed.

Add filter support to allow a user to make sure a BSP is present or
to remove BSPs that will not build.

Added the Init support because ARM BSPs do not contain the support.