RAP file support.
This is the initial addition of support for LZ77 RTEMS applications
in the RAP format.

The change add support for different formats to the loader. Each
format type is asked to check the file being loading and if
it matches the format's loader is called to load the file.

Support has been added to stream LZ77 files via the object file
cache.