Fix rela addend
If the bit 31 of info is 0, it means this is a reloc based on section.
In rtl-rap.c this line 'symvalue = (Elf_Word) symsect->base + addend'
show that if the bit 31 of info is 0, addend will be added to symvalue.
Then if the reloc type is rela and the reloc entry is based on section,
the addend should be assigned 0, otherwise error will be incured in the
reloc resolving function.